feat: optimize packaging
