+++ package.json
@@ -5,7 +5,8 @@
   "files": [
     "dist"
   ],
-  "module": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "main": "./dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
     "dev": "vite",
