+++ src/sunmao/app.ts
@@ -1,0 +1,12 @@
+const yaml = import.meta.glob("./dependencies/yaml/*.yaml", {
+  as: "raw",
+});
+
+export const dependencies = {
+  yaml: Object.keys(yaml).reduce<Record<string, string>>((prev, cur) => {
+    prev[cur.replace("./dependencies/yaml/", "").replace(".yaml", "")] = yaml[
+      cur
+    ] as unknown as string;
+    return prev;
+  }, {}),
+};
