+++ src/sunmao/lib.ts
@@ -99,15 +99,3 @@ export const libs: SunmaoLib[] = [
   },
 ];
 
-const yaml = import.meta.glob("./dependencies/yaml/*.yaml", {
-  as: "raw",
-});
-
-export const dependencies = {
-  yaml: Object.keys(yaml).reduce<Record<string, string>>((prev, cur) => {
-    prev[cur.replace("./dependencies/yaml/", "").replace(".yaml", "")] = yaml[
-      cur
-    ] as unknown as string;
-    return prev;
-  }, {}),
-};
